refactor types and server
