--- src/data/environment.js
@@ -1,3 +1,0 @@
-import dotenv from 'dotenv';
-
-export default dotenv.config();
