+++ src/infrastructure/readEnv.js
@@ -1,0 +1,7 @@
+import dotenv from "dotenv";
+
+const readEnviromentVariables = () => {
+    dotenv.config();
+};
+
+export default readEnviromentVariables;
