+++ src/server/schemaBuilder.js
@@ -1,0 +1,9 @@
+import typeDefs from "../types";
+import { resolvers } from "../resolvers/index.js";
+
+const schemaDefs = {
+  typeDefs,
+  resolvers,
+};
+
+export { schemaDefs };
